level made, main map, menus
pause menu and main menu made.

wavespawner, wave manager link up to automatically create a set of rounds based on the 'difficulty' variable.

map has nodes that can be travelled to, they load in a new level when clicked on.

when you finish all rounds of a level it takes you back to map.